requirements.txt
selenium
python-dotenv
pandas
lxml
pyodbc